Mock unit test
